engine: Don't rely on pydantic's copy as it is a shallow copy
The engine modifies the pipeline with arguments given by the source
but, because pydantic's copy is a shallow copy, pipeline would only
be changed once with the first source arguments.

The visible symptom was that the same manifest+pipeline pair would
give different results via CLI compared to webservice. Due to the
fact that webservice would pass full pipeline objects already.