destinations: Initial destinations support
In our current API, artifacts can be written  only on
the local filesystem. This is  quite  restrictive and
makes a lot of assumptions as for how API clients can
interact with ingestum.

Remove this limitation by introducing Destinations. Simply
put is the counter part of Locations. Ingestum clients can
now specify where the artifacts should be stored.

RM#1168