transformers: Exclude alternative language PubMed abstracts
PubMed XML includes <OtherAbstract> tag which can contain
alternative versions of the abstract. Including abstracts
in other languages, which we don't want.

Therefore, stick to <Abstract> if available, since it's
always in English [1]. And, stick to <OtherAbstract> but
only if it's in English.

If both are present, and in English, keep both.

[1] https://www.nlm.nih.gov/bsd/licensee/elements_descriptions.html

ENG-1423