transformers: Break down Biorxiv's get_publication method
Breaking the method into smaller parts will allow us to reuse the code to create two more transformers,
one that creates a biorxiv xml document from the source and another one that takes a biorxiv xml document to then transform it into a publication document.
The goal is to be able to save the raw data and process it as many times as we want without having to download it again.

INGEST-19