ingestum: Remove workers API
We need to re-think how we want to make Ingestum to scale-up.

This implementation was never really used in production  and
it introduced some  bugs  on how memory is handled,  e.g. it
keeps  all  results in memory even when these are not really
needed anymore, making Ingestum consume 2x memory.

INGEST-81