engine: Add alternative more memory-lightweight runner
The current runner has a major flaw,  it keeps all results
in memory. This is only needed  by ingestum-pipeline so it
can output the results to stdout more easily.

But, this is not fit for production where we only require
references to these results, not the actual results.

Therefore, let's add an alternative runner that's memory
conscious.

Use this in all production-level tools.

INGEST-81